# Improve similarities performance