@mihaiblidaru 2022-11-23
I just discovered that there are official benchmarks. I ran then using github workflows which was a terrible idea because the VMs they run on are veeeeeeeery sloooooow.

There's not much improvement when benchmarking the ml-100k dataset
**Before**
| [Movielens 100k](https://grouplens.org/datasets/movielens/100k)                                                                         |   RMSE |   MAE | Time    |
|:----------------------------------------------------------------------------------------------------------------------------------------|-------:|------:|:--------|
| [k-NN](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNBasic)                        |  0.98  | 0.774 | 0:00:17 |
| [Centered k-NN](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNWithMeans)           |  0.951 | 0.749 | 0:00:18 |
| [k-NN Baseline](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNBaseline)            |  0.931 | 0.733 | 0:00:21 |

**After**
| [Movielens 100k](https://grouplens.org/datasets/movielens/100k)                                                                         |   RMSE |   MAE | Time    |
|:----------------------------------------------------------------------------------------------------------------------------------------|-------:|------:|:--------|
| [k-NN](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNBasic)                        |  0.98  | 0.774 | 0:00:16 |
| [Centered k-NN](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNWithMeans)           |  0.951 | 0.749 | 0:00:17 |
| [k-NN Baseline](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNBaseline)            |  0.931 | 0.733 | 0:00:19 |

Using the ml-1m dataset my implementation is slower for some unknown reason and made me doubt the consistency of the performance of the github VMs.

**Before**
| [Movielens 1M](https://grouplens.org/datasets/movielens/1m)                                                                   |   RMSE |   MAE | Time    |
|:------------------------------------------------------------------------------------------------------------------------------|-------:|------:|:--------|
| [k-NN](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNBasic)              |  0.923 | 0.727 | 0:11:51 |
| [Centered k-NN](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNWithMeans) |  0.929 | 0.738 | 0:12:03 |
| [k-NN Baseline](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNBaseline)  |  0.895 | 0.706 | 0:11:36 |

**After**
| [Movielens 1M](https://grouplens.org/datasets/movielens/1m)                                                                   |   RMSE |   MAE | Time    |
|:------------------------------------------------------------------------------------------------------------------------------|-------:|------:|:--------|
| [k-NN](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNBasic)              |  0.923 | 0.727 | 0:13:31 |
| [Centered k-NN](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNWithMeans) |  0.929 | 0.738 | 0:13:28 |
| [k-NN Baseline](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNBaseline)  |  0.895 | 0.706 | 0:13:40 |

I ran the benchmarks again on a `c5.2xlarge` AWS EC2 instance and the benchmark results make more sense.
**Before**
| [Movielens 1M](https://grouplens.org/datasets/movielens/1m)                                                                   |   RMSE |   MAE | Time    |
|:------------------------------------------------------------------------------------------------------------------------------|-------:|------:|:--------|
| [k-NN](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNBasic) | 0.923 | 0.727 | 0:09:22 |

**After**
| [Movielens 1M](https://grouplens.org/datasets/movielens/1m)                                                                   |   RMSE |   MAE | Time    |
|:------------------------------------------------------------------------------------------------------------------------------|-------:|------:|:--------|
| [k-NN](https://surprise.readthedocs.io/en/stable/knn_inspired.html#surprise.prediction_algorithms.knns.KNNBasic) | 0.923 | 0.727 | 0:07:45 |

My guess is that the improvement is lower when running the full benchmark because most of the time is spent testing the algorithm and generating predictions rather than training it.